+++ Backend/UniversitySpace/src/main/java/ru/ruslan/repository/chat/ChatRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import ru.ruslan.entity.chat.Chat;
 
 @Repository
-public interface ChatRepository extends JpaRepository<Chat, Integer> {
-    Chat getByChatId(Integer chatId);
+public interface ChatRepository extends JpaRepository<Chat, Long> {
+    Chat getByChatId(Long chatId);
     Chat findByName(String name);
 }
